Move ADDED tests
