+++ src/test/java/org/opentripplanner/updater/trip/moduletests/delay/SkippedTest.java
@@ -46,7 +46,6 @@ public class SkippedTest {
 
     assertSuccess(env.applyTripUpdate(tripUpdate));
 
-
     final TimetableSnapshot snapshot = env.getTimetableSnapshot();
 
     // Original trip pattern
